Allow downloading projects from asset library
This commit adds support for downloading projects from the asset
library.  A new tab has been added to the main screen which allows
browsing, filtering and sorting assets.  Clicking on an asset listing
brings up some information on the asset and the option to download it.

This commit does not add support for asset icons and asset previews.
For now, all assets will display with the default icon.svg as their
icon.