Fix ok button not updating when file selected
Add a call to _update_ok_button when files are selected.

It appears that the line edit text changed signal was not detecting the
changes to the path edit or name edit nodes text vales. As such
_update_ok_button was not getting called when expected leading to the
ok button being disabled when a valid path is inputted.

Adding a call to _update_ok_button when a file is selected means it will
check if confirmation should be allowed when expected.