Handle edge cases and update desktop database after install
This commit handles a few edge cases for downloading files and
unzipping. I used the python3 method as the default for unzipping
because I think most systems should have python installed, since it is a
dependacy for a lot widely used applications.

I also added some lines to update the desktop database after installing.
This was added mostly because of my personal Hyprland configuration,
where it doesn't do this automatically, although on most DE's like Gnome
or Plasma doing this would be unnecessary.